Rename List -> LinkedList, with a List alias pointing to LinkedList.
